+++ src/WebJobs.Script/ExtensionBundle/ExtensionBundleManager.cs
@@ -252,45 +252,67 @@ namespace Microsoft.Azure.WebJobs.Script.ExtensionBundle
 
         internal string FindBestVersionMatch(VersionRange versionRange, IEnumerable<string> versions, string bundleId, FunctionsHostingConfigOptions configOption)
         {
+            int dotnetVersion = typeof(string).Assembly.GetName().Version.Major;
+
+            // Limit the version range for v4.x bundles on .NET 6 to be between 4.2.0 and 4.22.0
+            // Return original version range if not .NET 6 or not v4.x bundle
+            var effectiveVersionRange = GetAdjustedVersion(dotnetVersion, versionRange, bundleId);
+
+            // Check if there is a max version configured in hosting config for the current dotnet version and bundle major version
+            if (TryGetMaxBundleVersionFromHostingConfig(bundleId, versionRange.MinVersion.Major, dotnetVersion, configOption, out NuGetVersion maxBundleVersion))
+            {
+                effectiveVersionRange = new VersionRange(effectiveVersionRange.MinVersion, effectiveVersionRange.IsMinInclusive, maxBundleVersion, true);
+            }
+
             var bundleVersions = versions.Select(p =>
             {
                 var dirName = Path.GetFileName(p);
                 NuGetVersion.TryParse(dirName, out NuGetVersion version);
                 if (version != null)
                 {
-                    version = versionRange.Satisfies(version) ? version : null;
+                    version = effectiveVersionRange.Satisfies(version) ? version : null;
                 }
                 return version;
             }).Where(v => v != null).OrderByDescending(version => version.Version).ToList();
 
             var matchingVersion = ResolvePlatformReleaseChannelVersion(bundleVersions);
 
-            if (bundleId != ScriptConstants.DefaultExtensionBundleId)
-            {
-                return matchingVersion?.ToString();
-            }
-
-            // Check to see if there is a max bundle version set via hosting configuration, if yes then use that instead of the one
-            // available on VM or local machine. Only use MaximumBundleV3Version or MaximumBundleV4Version if the version configured
-            // by the customer resolved to version higher than the version set via hosting config.
-            if (!string.IsNullOrEmpty(configOption.MaximumBundleV3Version)
-                && matchingVersion?.Major == ScriptConstants.ExtensionBundleV3MajorVersion)
-            {
-                var maximumBundleV3Version = NuGetVersion.Parse(configOption.MaximumBundleV3Version);
-                matchingVersion = matchingVersion > maximumBundleV3Version ? maximumBundleV3Version : matchingVersion;
-                return matchingVersion?.ToString();
-            }
-
-            if (!string.IsNullOrEmpty(configOption.MaximumBundleV4Version)
-                && matchingVersion?.Major == ScriptConstants.ExtensionBundleV4MajorVersion)
-            {
-                var maximumBundleV4Version = NuGetVersion.Parse(configOption.MaximumBundleV4Version);
-                matchingVersion = matchingVersion > maximumBundleV4Version
-                                ? maximumBundleV4Version
-                                : matchingVersion;
-            }
-
             return matchingVersion?.ToString();
+        }
+
+        // Applies bundle version limits for .NET 6, using the default bundle if referencing a v4 major bundle version.
+        private VersionRange GetAdjustedVersion(int dotnetVersion, VersionRange versionRange, string bundleId)
+        {
+            if (dotnetVersion != 6
+                || !string.Equals(bundleId, ScriptConstants.DefaultExtensionBundleId, StringComparison.OrdinalIgnoreCase)
+                || versionRange.MinVersion.Major != ScriptConstants.ExtensionBundleV4MajorVersion)
+            {
+                return versionRange;
+            }
+
+            var effectiveMinVersion = new NuGetVersion(ScriptConstants.Net6MinimunV4BundleVersion);
+            var effectiveMaxVersion = new NuGetVersion(ScriptConstants.Net6MaximumV4BundleVersion);
+
+            var minVersion = versionRange.MinVersion < effectiveMinVersion ? effectiveMinVersion : versionRange.MinVersion;
+            var maxVersion = versionRange.MaxVersion > effectiveMaxVersion ? effectiveMaxVersion : versionRange.MaxVersion;
+
+            // Use the original inclusion values if it is within the min or max range.
+            bool isMinInclusive = versionRange.MinVersion > effectiveMinVersion && versionRange.IsMinInclusive;
+            bool isMaxInclusive = versionRange.MaxVersion < effectiveMaxVersion && versionRange.IsMaxInclusive;
+
+            versionRange = new VersionRange(minVersion, isMinInclusive, maxVersion, isMaxInclusive);
+            return versionRange;
+        }
+
+        private bool TryGetMaxBundleVersionFromHostingConfig(string bundleId, int bundleVersion, int dotnetVersion, FunctionsHostingConfigOptions configOption, out NuGetVersion maxVersion)
+        {
+            maxVersion = null;
+            string hostingConfig = $"Net{dotnetVersion}MaximumBundleV{bundleVersion}Version";
+            string hostingConfigValue = configOption.GetFeature(hostingConfig);
+
+            return string.Equals(bundleId, ScriptConstants.DefaultExtensionBundleId, StringComparison.OrdinalIgnoreCase)
+                && !string.IsNullOrEmpty(hostingConfigValue)
+                && NuGetVersion.TryParse(hostingConfigValue, out maxVersion);
         }
 
         private NuGetVersion ResolvePlatformReleaseChannelVersion(IList<NuGetVersion> orderedByDescBundles) => _platformReleaseChannel.ToUpper() switch