+++ release_notes.md
@@ -3,4 +3,5 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
-- Add JitTrace files for v4.43 (#11277)
+Add JitTrace files for v4.43 (#11277)
+Cap v4.x extension bundle versions on .NET 6 to the range (>4.2.0, <4.22.0) by default (#11289)